+++ services/teacher.rb
@@ -1,0 +1,17 @@
+require_relative '../services/person'
+# inherits from person
+class Teacher < Person
+  attr_accessor :specialization
+
+  # initialize
+  def initialize(name, age, specialization)
+    super(name, age)
+    @specialization = specialization
+  end
+
+  # can_use_services
+  def can_use_services?
+    true
+  end
+  private :can_use_services?
+end
